abstraccion entre menu y controlador
