+++ options/game.py
@@ -1,4 +1,7 @@
 from enum import Enum
+import os
+
+from options.menu import ExitGameException, MenuOption
 
 ABANDON_GAME_OPT = "0"
 ASK_CLUE_OPT = "1"
@@ -23,9 +26,24 @@ class GameOpt(Enum):
     def execute(self, juego):
         match self:
             case GameOpt.ABANDON_GAME_OPT:
-                juego.game_abandoned()   # No se utiliza
+                if self.ask_abandon_confirmation():
+                    raise ExitGameException
             case GameOpt.ASK_CLUE:
                 juego.give_clue()
             case GameOpt.ASK_HELP:
                 juego.give_help()
 
+    def ask_abandon_confirmation(self):
+        if self == GameOpt.ABANDON_GAME_OPT:
+            print("\n")
+            while True:
+                print("¿Estas seguro de que deseas abandonar la partida?")
+                print("0. Si")
+                print("1. No\n")
+                inp = input("- ").strip()
+                os.system('clear')
+                if inp == "0":
+                    return True
+                if inp == "1":
+                    return False
+                MenuOption.show_incorrect_option_message()
