+++ options/menu.py
@@ -1,4 +1,5 @@
 from enum import Enum
+import os
 import controller
 START_GAME_OPT = "1"
 SELECT_DIFFICULTY_OPT = "2"
@@ -12,7 +13,7 @@ class MenuOption(Enum):
     EXIT = 4
     
     def show_incorrect_option_message():
-        print("Opcion incorrecta")
+        print("Opcion incorrecta, vuelve a intentarlo\n")
 
     def from_input(inp):
         if inp == START_GAME_OPT:
@@ -33,9 +34,33 @@ class MenuOption(Enum):
             case MenuOption.SELECT_DIFFICULTY:
                 game_controller.update_difficulty()
             case MenuOption.RULES:
-                controller.GameController.show_rules()
+                game_controller.show_rules()
             case MenuOption.EXIT:
-                print("¡Gracias por jugar!")
+                if self.ask_exit_confirmation():
+                    print("\n¡Gracias por jugar!\n")
+                    raise ExitGameException 
             case _:
                 print("Opcion incorrecta")
 
+
+    def ask_exit_confirmation(self):
+        if self == MenuOption.EXIT:
+            print("\n")
+            os.system('clear')
+            while True:
+                print("¿Estas seguro de que deseas salir del juego?")
+                print("0. Si")
+                print("1. No\n")
+                inp = input("- ").strip()
+                os.system('clear')
+                if inp == "0":
+                    return True
+                if inp == "1":
+                    return False
+                MenuOption.show_incorrect_option_message()
+
+        return False
+    
+
+class ExitGameException(Exception):
+    pass
